toolkit/components/alerts/test/test_alerts_noobserve.html

<!DOCTYPE HTML>
<!-- Any copyright is dedicated to the Public Domain.
     http://creativecommons.org/publicdomain/zero/1.0/ -->
<html>
<head>
  <title>Test for Alerts Service</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="head.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>

<body>
<p id="display"></p>

<br>Alerts service, without observer "asynchronous" case.
<br>
<br>A notification should soon appear somewhere.
<br>If there has been no crash when the notification (later) disappears, assume all is good.

<pre id="test">
<script class="testbody" type="text/javascript">
const Cc = SpecialPowers.Cc;
const Ci = SpecialPowers.Ci;

const chromeScript = SpecialPowers.loadChromeScript(_ => {
  /* eslint-env mozilla/chrome-script */
  const {setTimeout} = ChromeUtils.import("resource://gre/modules/Timer.jsm");

  function anyXULAlertsVisible() {
    var windows = Services.wm.getEnumerator("alert:alert");
    return windows.hasMoreElements();
  }

  addMessageListener("waitForAlerts", function waitForAlerts() {
    if (anyXULAlertsVisible()) {
      setTimeout(waitForAlerts, 1000);
    } else {
      sendAsyncMessage("waitedForAlerts");
    }
  });
});

function waitForAlertsThenFinish() {
  var resolved = new Promise(resolve => {
    chromeScript.addMessageListener("waitedForAlerts", function waitedForAlerts() {
      chromeScript.removeMessageListener("waitedForAlerts", waitedForAlerts);
      ok(true, "Alert disappeared.");
      resolve();
    });
  });
  chromeScript.sendAsyncMessage("waitForAlerts");

  return resolved;
}

add_task(async function testNoObserve() {
  if (!("@mozilla.org/alerts-service;1" in Cc)) {
    todo(false, "Alerts service does not exist in this application");
  } else {
    ok(true, "Alerts service exists in this application");

    var notifier;
    try {
      notifier = Cc["@mozilla.org/alerts-service;1"].
                 getService(Ci.nsIAlertsService);
      ok(true, "Alerts service is available");
    } catch (ex) {
      todo(false, "Alerts service is not available.", ex);
    }

    if (notifier) {
      try {
        notifier.showAlertNotification(null, "Notification test",
                                       "This notification has no observer");
        ok(true, "showAlertNotification() succeeded");

        await new Promise(resolve => {
          // Add delay to wait for `alert:alert` window creation in case this is a XUL alert.
          setTimeout(resolve, 1000);
        }).then(async () => {
          await waitForAlertsThenFinish();
        });
      } catch (ex) {
        todo(false, "showAlertNotification() failed.", ex);
      }
    }
  }
});

SimpleTest.requestFlakyTimeout("`alert:alert` window creation and deletion " +
                               "lifetime may not be synchronous with the alert object.");
</script>
</pre>
</body>
</html>
